Basic Python binding.
 * Python binding with Boost::Python
 * so far does not use Boost.Numpy
 * Python module `qca`
 * updated README.md

+++ src/updateVersion.hpp.sh
@@ -1,7 +1,13 @@
 #! /bin/bash
 
-GIT_PROGRAM_VERSION=`git describe --always | tr -d "\n"`
-sed 's/${GIT_PROGRAM_VERSION}/'$GIT_PROGRAM_VERSION'/g' version.hpp.in > version.hpp.tmp
+PROGRAM_VERSION=`git describe --always --dirty | tr -d "\n"`
+if [[ "$PROGRAM_VERSION" == "" ]]; then
+    PROGRAM_VERSION="unknown"
+fi
+if [[ -f version.hpp && "$PROGRAM_VERSION" == "unknown" ]]; then
+    exit 0;
+fi
+sed 's/${PROGRAM_VERSION}/'$PROGRAM_VERSION'/g' version.hpp.in > version.hpp.tmp
 if ( diff version.hpp version.hpp.tmp > /dev/null 2>&1 ); then 
    rm version.hpp.tmp
 else
